Moved MssParser into mss/parser
